requirements.txt
coverage==4.5.1
Django==1.11.10
psycopg2==2.7.4
PyGithub==1.36
PyJWT==1.5.3
pytz==2018.3
